requirements.txt
attrs @ file:///private/tmp/python-attrs-20231125-8592-pup5dc/attrs-23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chess==1.10.0
click @ file:///private/tmp/python-click-20231211-45440-9p5m4p/click-8.1.7
contourpy==1.1.0
cryptography==42.0.5
cycler==0.11.0
firebase-admin==6.5.0
Flask-Cors==4.0.0
fonttools==4.42.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
imbalanced-learn==0.11.0
imblearn==0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2 @ file:///private/tmp/python-jinja-20231112-5160-1o1crd6/Jinja2-3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe @ file:///private/tmp/python-markupsafe-20231211-21370-34xmob/MarkupSafe-2.1.3
matplotlib==3.7.2
msgpack==1.0.8
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pathspec @ file:///private/tmp/python-pathspec-20231212-4674-kki6wg/pathspec-0.12.1
pdf2image==1.16.3
Pillow==10.0.0
platformdirs @ file:///private/tmp/python-platformdirs-20231204-5228-f28h82/platformdirs-4.1.0
praw==7.7.1
prawcore==2.4.0
proto-plus==1.23.0
protobuf==4.25.3
psutil @ file:///private/tmp/python-psutil-20231015-5276-sjz8wa/psutil-5.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo @ file:///private/tmp/py3cairo-20231211-32963-tvpe2z/pycairo-1.25.1
pycparser==2.22
Pygments @ file:///private/tmp/pygments-20231121-5079-e7uuzt/pygments-2.17.2
PyGObject==3.46.0
PyJWT==2.8.0
pyparsing==3.1.2
python-chess==1.999
python-dateutil==2.8.2
pytz==2023.3
PyYAML @ file:///private/tmp/pyyaml-20231018-5457-utbrds/PyYAML-6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scikit-multilearn==0.2.0
scipy==1.11.3
seaborn==0.12.2
six==1.16.0
soupsieve==2.5
threadpoolctl==3.2.0
tweepy==4.14.0
tzdata==2023.3
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.2
